Change of file name and test descriptions
